Crate libportal

libportal, the portal gun getting oxedized, the Rust wrapper for the XDG portals DBus interfaces.

use libportal::desktop::screenshot::{ScreenshotProxy, PickColorOptions};
use libportal::WindowIdentifier;

fn main() -> zbus::fdo::Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = ScreenshotProxy::new(&connection)?;
    let request = proxy.pick_color(WindowIdentifier::default(), PickColorOptions::default())?;
    Ok(())
}

Re-exports

pub use zbus;

pub use zvariant;

Modules

desktop
Implementation of the various portals under /org/freedesktop/portal/desktop

documents
Implementation of the various portals under /org/freedesktop/portal/documents

flatpak
Implementation of the various portals under /org/freedesktop/portal/Flatpak

Structs

BasicResponse

NString
Null terminated string

RequestProxy
The Request interface is shared by all portal interfaces. When a portal method is called, the reply includes a handle (i.e. object path) for a Request object, which will stay alive for the duration of the user interaction related to the method call.

SessionProxy
The Session interface is shared by all portal interfaces that involve long lived sessions. When a method that creates a session is called, if successful, the reply will include a session handle (i.e. object path) for a Session object, which will stay alive for the duration of the session.

WindowIdentifier

Enums

ResponseError

Type Definitions

Response