Module libportal::desktop::inhibit

Interact with org.freedesktop.portal.Inhibit interface.

Structs

CreateMonitorOptions

InhibitOptions

InhibitProxy
The interface lets sandboxed applications inhibit the user session from ending, suspending, idling or getting switched away.