Module libportal::desktop::screencast

Interact with org.freedesktop.portal.ScreenCast interface.

Examples

How to create a screen cast session & start it. The portal is currently useless without pipewire & rust support.

use libportal::desktop::screencast::{
    CreateSession, CreateSessionOptions, CursorMode, ScreenCastProxy, SelectSourcesOptions,
    SourceType, StartCastOptions, Streams,
};
use libportal::{BasicResponse as Basic, HandleToken, RequestProxy, Response, WindowIdentifier};
use zbus::{self, fdo::Result};
use zvariant::ObjectPath;
use enumflags2::BitFlags;
use std::convert::TryFrom;

fn select_sources(
    session_handle: ObjectPath,
    proxy: &ScreenCastProxy,
    connection: &zbus::Connection,
) -> Result<()> {
    let request_handle = proxy.select_sources(
        session_handle.clone(),
        SelectSourcesOptions::default()
            .multiple(true)
           .cursor_mode(BitFlags::from(CursorMode::Metadata))
            .types(SourceType::Monitor | SourceType::Window),
    )?;

    let request = RequestProxy::new(&connection, &request_handle)?;
    request.on_response(move |response: Response<Basic>| {
        if response.is_ok() {
            start_cast(session_handle, proxy, connection).unwrap();
        }
    })?;
    Ok(())
}

fn start_cast(
    session_handle: ObjectPath,
    proxy: &ScreenCastProxy,
    connection: &zbus::Connection,
) -> Result<()> {
    let request_handle = proxy.start(
        session_handle,
        WindowIdentifier::default(),
        StartCastOptions::default(),
    )?;
    let request = RequestProxy::new(&connection, &request_handle)?;
    request.on_response(move |r: Response<Streams>| {
        r.unwrap().streams().iter().for_each(|stream| {
            println!("{}", stream.pipewire_node_id());
            println!("{:#?}", stream.properties());
        });
    })?;
    Ok(())
}

fn main() -> Result<()> {
    let connection = zbus::Connection::new_session()?;
    let proxy = ScreenCastProxy::new(&connection)?;

    let request_handle =
        proxy.create_session(CreateSessionOptions::default().session_handle_token(HandleToken::try_from("token").unwrap()))?;
    let request = RequestProxy::new(&connection, &request_handle)?;

    request.on_response(|r: Response<CreateSession>| {
        match r {
            Ok(session) => select_sources(session.handle(), &proxy, &connection).unwrap(),
            Err(_) => println!("hello!"),
        };
    })?;
    Ok(())
}

Structs

CreateSession

CreateSessionOptions
Specified options on a create a screencast session request.

ScreenCastProxy
The interface lets sandboxed applications create screen cast sessions.

SelectSourcesOptions
Specified options on a select sources request.

StartCastOptions
Specified options on a start screencast request.

Stream

StreamProperties

Streams

Enums

CursorMode

SourceType