Module libportal::flatpak

Implementation of the various portals under /org/freedesktop/portal/Flatpak

Modules

update_monitor

Structs

CreateMonitorOptions

FlatpakProxy
The interface exposes some interactions with Flatpak on the host to the sandbox. For example, it allows you to restart the applications or start a more sandboxed instance.

SpawnOptions